move schema allocator to public header
