+++ r/src/pointers.c
@@ -26,7 +26,7 @@
 // More reliable way to stringify intptr_t on Windows using C++
 void intptr_as_string(intptr_t ptr_int, char* buf);
 
-SEXP nanoarrow_c_allocate_schema(void) { return schema_owning_xptr(); }
+SEXP nanoarrow_c_allocate_schema(void) { return nanoarrow_schema_owning_xptr(); }
 
 SEXP nanoarrow_c_allocate_array(void) { return array_owning_xptr(); }
